# Performance improvement: Preserve order of sources