# Description

Currently, the dependencies are in reverse order of use when it needs to do the `cleanupSources`. There is quite some time spent doing so.

While doing some experiments on a project on mine I found out that it'd perform faster if the dependencies order didn't need to be built in reverse order. The way to achieve this is by doing two things:
1. On `prepareSources`, make `target._sources` point to the tail instead of the head
2. On `cleanupSources`, walk from tail to head **while** dropping old dependencies and reverting to the rollback node

The advantage of doing so is that, if there are no nodes to drop (nodes with `version: -1`) then we're just walking the list backwards without setting any properties.

New nodes are added directly to the tail, and the new node becomes the tail. If old nodes were re-used, then adding new nodes to the tail is cheap and also means that `cleanupSource` only needs to walk backwards until it finds the head without setting properties (that's besides rollback node).

I commented the code with examples so it'd also be easier to understand for new people looking at it 😅 (it took me a bit to grasp how it works).

## Benchmarks

Benchmarked on `v8` binary, Node.js (diff V8 version), and SpiderMonkey (`sm -m` as module). I'm unable to benchmark on JavaScriptCore (can't find a built [trusted] binary).

```diff
v8 (10.9.84)
====================================
Signals  Core - Simple x 25,555,234 ops/sec ±1.07% (487177 runs sampled)
Signals  Core - Unchanged sources x 53,590 ops/sec ±0.47% (1042 runs sampled)
Signals  Core - Unchanged sources + Effect x 50,580 ops/sec ±0.55% (983 runs sampled)
Signals  Core - Conditional Sources x 78,645 ops/sec ±0.57% (1529 runs sampled)
Signals  Core - Conditional Sources + Effect x 73,532 ops/sec ±0.44% (1429 runs sampled)

! Signals Local - Simple x 26,272,524 ops/sec ±0.19% (498317 runs sampled)
! Signals Local - Unchanged sources x 57,449 ops/sec ±0.49% (1117 runs sampled)
! Signals Local - Unchanged sources + Effect x 55,694 ops/sec ±0.53% (1083 runs sampled)
! Signals Local - Conditional Sources x 83,705 ops/sec ±0.50% (1628 runs sampled)
! Signals Local - Conditional Sources + Effect x 80,809 ops/sec ±0.55% (1571 runs sampled)

Node (18.11.0, V8: 10.2.154.15-node.12)
====================================
Signals  Core - Simple x 25,533,845 ops/sec ±0.09% (484112 runs sampled)
Signals  Core - Unchanged sources x 47,989 ops/sec ±0.57% (933 runs sampled)
Signals  Core - Unchanged sources + Effect x 49,957 ops/sec ±0.50% (972 runs sampled)
Signals  Core - Conditional Sources x 80,676 ops/sec ±0.60% (1569 runs sampled)
Signals  Core - Conditional Sources + Effect x 78,402 ops/sec ±0.58% (1525 runs sampled)

! Signals Local - Simple x 25,713,380 ops/sec ±0.09% (487306 runs sampled)
! Signals Local - Unchanged sources x 50,811 ops/sec ±0.62% (987 runs sampled)
! Signals Local - Unchanged sources + Effect x 52,418 ops/sec ±0.54% (1019 runs sampled)
! Signals Local - Conditional Sources x 92,738 ops/sec ±0.57% (1804 runs sampled)
! Signals Local - Conditional Sources + Effect x 81,292 ops/sec ±0.53% (1581 runs sampled)

SpiderMonkey (JavaScript-C107.0)
====================================
Signals  Core - Simple x 11,124,584 ops/sec ±0.11% (214318 runs sampled)
Signals  Core - Unchanged sources x 30,535 ops/sec ±0.61% (594 runs sampled)
Signals  Core - Unchanged sources + Effect x 21,317 ops/sec ±0.54% (415 runs sampled)
Signals  Core - Conditional Sources x 51,024 ops/sec ±0.46% (992 runs sampled)
Signals  Core - Conditional Sources + Effect x 43,003 ops/sec ±0.47% (837 runs sampled)

! Signals Local - Simple x 11,198,532 ops/sec ±0.10% (215715 runs sampled)
! Signals Local - Unchanged sources x 32,117 ops/sec ±0.53% (625 runs sampled)
! Signals Local - Unchanged sources + Effect x 23,483 ops/sec ±0.60% (457 runs sampled)
! Signals Local - Conditional Sources x 58,449 ops/sec ±0.51% (1137 runs sampled)
! Signals Local - Conditional Sources + Effect x 46,252 ops/sec ±0.49% (900 runs sampled)
```

And one un-proper naive benchmark 😅 
```ts
const count = s[250];
const compt = computed(() => {
	for (let i = 0; i < 1000; i++) {
		const sig = s[i];
		sig.value;
	}

	s1.value; // s[1] and so on (check already seen node)
	s2.value;
	s3.value;
	s4.value;
	s5.value;
	s6.value;
	s7.value;
	s8.value;

	return count.value * 2;
});

console.time("Naive:  Core");
for (let i = 0; i < 200000; i++) {
	count.value++;
	compt.value;
}
console.timeEnd("Naive:  Core");
```

```diff
console.timeEnd: Naive:  Core, 3886.450000
console.timeEnd: Naive: Local, 3515.682000
```